Add a restart to disable fp exceptions.
Implemented for darwin-arm64 and linux-x86-64.

+++ src/runtime/x86-64-linux-os.h
@@ -4,7 +4,7 @@
 typedef ucontext_t os_context_t;
 typedef long os_context_register_t;
 
-unsigned long os_context_fp_control(os_context_t *context);
+unsigned int os_context_fp_control(os_context_t *context);
 #define RESTORE_FP_CONTROL_FROM_CONTEXT
 void os_restore_fp_control(os_context_t *context);
